Cleaned up canex service interaction
