--- projects/canpay-lib/src/lib/canpay-payment-qr/contract-one.json
@@ -1,211 +1,0 @@
-[{
-	"constant": true,
-	"inputs": [],
-	"name": "name",
-	"outputs": [{
-		"name": "",
-		"type": "string"
-	}],
-	"payable": false,
-	"stateMutability": "view",
-	"type": "function"
-}, {
-	"constant": false,
-	"inputs": [{
-		"name": "_spender",
-		"type": "address"
-	}, {
-		"name": "_value",
-		"type": "uint256"
-	}],
-	"name": "approve",
-	"outputs": [{
-		"name": "",
-		"type": "bool"
-	}],
-	"payable": false,
-	"stateMutability": "nonpayable",
-	"type": "function"
-}, {
-	"constant": true,
-	"inputs": [],
-	"name": "totalSupply",
-	"outputs": [{
-		"name": "",
-		"type": "uint256"
-	}],
-	"payable": false,
-	"stateMutability": "view",
-	"type": "function"
-}, {
-	"constant": false,
-	"inputs": [{
-		"name": "_from",
-		"type": "address"
-	}, {
-		"name": "_to",
-		"type": "address"
-	}, {
-		"name": "_value",
-		"type": "uint256"
-	}],
-	"name": "transferFrom",
-	"outputs": [{
-		"name": "",
-		"type": "bool"
-	}],
-	"payable": false,
-	"stateMutability": "nonpayable",
-	"type": "function"
-}, {
-	"constant": true,
-	"inputs": [{
-		"name": "",
-		"type": "address"
-	}],
-	"name": "balances",
-	"outputs": [{
-		"name": "",
-		"type": "uint256"
-	}],
-	"payable": false,
-	"stateMutability": "view",
-	"type": "function"
-}, {
-	"constant": true,
-	"inputs": [],
-	"name": "decimals",
-	"outputs": [{
-		"name": "",
-		"type": "uint256"
-	}],
-	"payable": false,
-	"stateMutability": "view",
-	"type": "function"
-}, {
-	"constant": true,
-	"inputs": [{
-		"name": "",
-		"type": "address"
-	}, {
-		"name": "",
-		"type": "address"
-	}],
-	"name": "allowed",
-	"outputs": [{
-		"name": "",
-		"type": "uint256"
-	}],
-	"payable": false,
-	"stateMutability": "view",
-	"type": "function"
-}, {
-	"constant": true,
-	"inputs": [{
-		"name": "_owner",
-		"type": "address"
-	}],
-	"name": "balanceOf",
-	"outputs": [{
-		"name": "",
-		"type": "uint256"
-	}],
-	"payable": false,
-	"stateMutability": "view",
-	"type": "function"
-}, {
-	"constant": true,
-	"inputs": [],
-	"name": "totalTokens",
-	"outputs": [{
-		"name": "",
-		"type": "uint256"
-	}],
-	"payable": false,
-	"stateMutability": "view",
-	"type": "function"
-}, {
-	"constant": true,
-	"inputs": [],
-	"name": "symbol",
-	"outputs": [{
-		"name": "",
-		"type": "string"
-	}],
-	"payable": false,
-	"stateMutability": "view",
-	"type": "function"
-}, {
-	"constant": false,
-	"inputs": [{
-		"name": "_to",
-		"type": "address"
-	}, {
-		"name": "_value",
-		"type": "uint256"
-	}],
-	"name": "transfer",
-	"outputs": [{
-		"name": "",
-		"type": "bool"
-	}],
-	"payable": false,
-	"stateMutability": "nonpayable",
-	"type": "function"
-}, {
-	"constant": true,
-	"inputs": [{
-		"name": "_owner",
-		"type": "address"
-	}, {
-		"name": "_spender",
-		"type": "address"
-	}],
-	"name": "allowance",
-	"outputs": [{
-		"name": "remaining",
-		"type": "uint256"
-	}],
-	"payable": false,
-	"stateMutability": "view",
-	"type": "function"
-}, {
-	"inputs": [],
-	"payable": false,
-	"stateMutability": "nonpayable",
-	"type": "constructor"
-}, {
-	"anonymous": false,
-	"inputs": [{
-		"indexed": true,
-		"name": "from",
-		"type": "address"
-	}, {
-		"indexed": true,
-		"name": "to",
-		"type": "address"
-	}, {
-		"indexed": false,
-		"name": "value",
-		"type": "uint256"
-	}],
-	"name": "Transfer",
-	"type": "event"
-}, {
-	"anonymous": false,
-	"inputs": [{
-		"indexed": true,
-		"name": "owner",
-		"type": "address"
-	}, {
-		"indexed": true,
-		"name": "spender",
-		"type": "address"
-	}, {
-		"indexed": false,
-		"name": "value",
-		"type": "uint256"
-	}],
-	"name": "Approval",
-	"type": "event"
-}]
