+++ projects/canpay-lib/src/lib/canpay-payment-qr/result.component.html
@@ -36,8 +36,6 @@
   <span style="position: absolute; right: 6%" *ngIf="!metamaskpayment">Metamask Not Supported</span>
 
   <div class="tip-button float-right" (click)="metamask()" *ngIf="metamaskpayment"> </div>
-
-
   <button class="btn btn-outline btn-white-trans btn_res payment-erc text-center" (click)="cancel()">
     Cancel</button>
   <div class="spinner">
