+++ projects/canpay-lib/src/lib/canpay.module.ts
@@ -10,22 +10,15 @@ import { ClipboardModule } from 'ngx-clipboard';
 
 import { ResizeService } from '../lib/services/resize.service';
 import { BancorWcComponent } from './bancor-wc/bancor-wc.component';
-import { FormDataService } from './canpay-data/formData.service';
 import { PaymentERCDetailsComponent } from './canpay-erc20-listing/payment-erc20.component';
-import { PaymentDetailsServiceERC } from './canpay-erc20-listing/payment-erc20.service';
 import { LoaderComponent } from './canpay-loader/loading-status.component';
 import { CanpayModalComponent } from './canpay-modal/canpay-modal.component';
 import { OrderComponent } from './canpay-order/order.component';
-import { OrderService } from './canpay-order/order.service';
 import { CompleteDetailsComponent } from './canpay-payment-complete/complete.component';
-import { CompleteDetailsService } from './canpay-payment-complete/complete.service';
 import { PaymentDetailsComponent } from './canpay-payment-details/payment-details.component';
-import { PaymentDetailsService } from './canpay-payment-details/payment-details.service';
 import { ErrorComponent } from './canpay-payment-error/error.component';
 import { ResultDetailsComponent } from './canpay-payment-qr/result.component';
-import { ResultService } from './canpay-payment-qr/result.service';
 import { StagingDetailsComponent } from './canpay-payment-staging/staging.component';
-import { StagingDetailsService } from './canpay-payment-staging/staging.service';
 import { CanpayWizardComponent } from './canpay-wizard/canpay-wizard.component';
 import { CommaSepNumPipe } from './comma-sep-num.pipe';
 import { InputAmountComponent } from './input-amount/input-amount.component';
@@ -39,9 +32,11 @@ import {
 } from './payment-authorisation/payment-authorisation.component';
 import { PaymentComponent } from './payment/payment.component';
 import { ProcessComponent } from './process/process.component';
+import { CanexService } from './services/canex.service';
 import { CanPayService } from './services/canpay.service';
 import { CanYaCoinEthService } from './services/canyacoin-eth.service';
 import { EthService } from './services/eth.service';
+import { FormDataService } from './services/formData.service';
 
 const COMPONENTS = [
   CanpayModalComponent,
@@ -69,8 +64,7 @@ const COMPONENTS = [
 
 const PROVIDERS = [
   EthService, CanYaCoinEthService, CanPayService, FormDataService,
-  OrderService, ResizeService, StagingDetailsService, PaymentDetailsServiceERC,
-  CompleteDetailsService, PaymentDetailsService, ResultService
+  ResizeService, CanexService
 ];
 
 @NgModule({
