+++ src/app/can-pay-example/can-pay-example.component.ts
@@ -27,7 +27,7 @@ export class CanPayExampleComponent {
     maxAmount: 50000, // Default is 'No Maximum'
 
     // CanExchange
-    destinationAddress: environment.contracts.testAccount, // account to credit from CanEx (can be overriden)
+    destinationAddress: undefined, // account to credit from CanEx (can be overriden)
     userEmail: '<your email>', // email to which the canexchange confirmation should be sent
     disableCanEx: false,
     // Actions
