+++ src/environments/environment.prod.ts
@@ -1,3 +1,10 @@
 export const environment = {
-  production: true
+  production: false,
+  contracts: {
+    canYaDao: '0xf558DBf7e2F1081D0331Dc0e025c2ECca12129d1',
+    canYaCoin: '0x1d462414fe14cf489c7a21cac78509f4bf8cd7c0',
+    useTestNet: false,
+    testAccount: '0xf558DBf7e2F1081D0331Dc0e025c2ECca12129d1'
+  },
+  ethNetId: 1524196056249
 };
