+++ src/environments/environment.ts
@@ -6,10 +6,10 @@
 export const environment = {
   production: false,
   contracts: {
-    canYaDao: '0xf558DBf7e2F1081D0331Dc0e025c2ECca12129d1',
-    canYaCoin: '0x1d462414fe14cf489c7a21cac78509f4bf8cd7c0',
-    useTestNet: false,
-    testAccount: '0xf558DBf7e2F1081D0331Dc0e025c2ECca12129d1'
+    canYaDao: '0x17b4ae55a5b0b6c10b0f4bae2d75a4e83de41709',
+    canYaCoin: '0x1b04ae0cbc58f813373ce2a854b97785762bdd34',
+    useTestNet: true,
+    testAccount: '0xA766743bD02AA07f5E5a7509F038028E1DEd8186'
   },
   ethNetId: 1524196056249
 };
